refactor: organized directory structure
