+++ app/src/main/java/com/example/b07demosummer2024/activities/MainActivity.java
@@ -1,24 +1,14 @@
-package com.example.b07demosummer2024;
+package com.example.b07demosummer2024.activities;
 
 import android.os.Bundle;
 
-import androidx.activity.EdgeToEdge;
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
 
-import android.util.Log;
-import android.widget.Button;
-import android.view.View;
-import android.content.Intent;
-
-import com.google.firebase.database.FirebaseDatabase;
-
-import java.util.*;
+import com.example.b07demosummer2024.utilities.Preferences;
+import com.example.b07demosummer2024.R;
+import com.example.b07demosummer2024.fragments.LoginFragment;
 
 public class MainActivity extends AppCompatActivity {
