+++ app/src/main/java/com/example/b07demosummer2024/fragments/RecyclerViewFragment.java
@@ -1,4 +1,4 @@
-package com.example.b07demosummer2024;
+package com.example.b07demosummer2024.fragments;
 
 import android.os.Bundle;
 import android.util.Log;
@@ -15,12 +15,14 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
-import com.google.firebase.database.DataSnapshot;
+
+import com.example.b07demosummer2024.utilities.Database;
+import com.example.b07demosummer2024.models.Item;
+import com.example.b07demosummer2024.utilities.ItemAdapter;
+import com.example.b07demosummer2024.R;
+import com.example.b07demosummer2024.interfaces.RecyclerViewInterface;
 import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.Query;
-import com.google.firebase.database.ValueEventListener;
+
 import java.util.ArrayList;
 import java.util.List;
