+++ app/src/main/java/com/example/b07demosummer2024/interfaces/LoginMVP.java
@@ -1,4 +1,4 @@
-package com.example.b07demosummer2024;
+package com.example.b07demosummer2024.interfaces;
 
 public interface LoginMVP
 {
