+++ app/src/main/java/com/example/b07demosummer2024/models/Media.java
@@ -1,4 +1,4 @@
-package com.example.b07demosummer2024;
+package com.example.b07demosummer2024.models;
 
 import java.util.List;
