+++ app/src/main/java/com/example/b07demosummer2024/unused/RecyclerViewStaticFragment.java
@@ -1,5 +1,5 @@
 // CURRENTLY NOT USED
-package com.example.b07demosummer2024;
+package com.example.b07demosummer2024.unused;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -11,6 +11,10 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.b07demosummer2024.models.Item;
+import com.example.b07demosummer2024.utilities.ItemAdapter;
+import com.example.b07demosummer2024.R;
 
 import java.util.ArrayList;
 import java.util.List;
