+++ app/src/main/java/com/example/b07demosummer2024/unused/ScrollerFragment.java
@@ -1,15 +1,16 @@
 //CURRENTLY NOT USED
-package com.example.b07demosummer2024;
+package com.example.b07demosummer2024.unused;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ScrollView;
 import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+
+import com.example.b07demosummer2024.R;
 
 public class ScrollerFragment extends Fragment {
     @Nullable
