+++ app/src/main/java/com/example/b07demosummer2024/unused/SpinnerFragment.java
@@ -1,5 +1,5 @@
 //CURRENTLY NOT USED
-package com.example.b07demosummer2024;
+package com.example.b07demosummer2024.unused;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,6 +10,8 @@ import android.widget.Spinner;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+
+import com.example.b07demosummer2024.R;
 
 public class SpinnerFragment extends Fragment {
     @Nullable
