+++ app/src/main/res/layout/activity_home_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeFragment">
+    tools:context=".unused.HomeFragment">
 
     <Button
         android:id="@+id/buttonRecyclerView"
