+++ app/src/main/res/layout/fragment_add_item.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddItemFragment">
+    tools:context=".fragments.AddItemFragment">
 
     <EditText
         android:id="@+id/editTextTitle"
