+++ app/src/main/res/layout/fragment_login_popup.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginPopup">
+    tools:context=".fragments.LoginPopup">
 
     <EditText
         android:id="@+id/editTextUsername"
