+++ app/src/main/res/layout/fragment_manage_items.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ManageItemsFragment">
+    tools:context=".fragments.ManageItemsFragment">
 
     <Button
         android:id="@+id/buttonAddItem"
