+++ app/src/main/res/layout/fragment_scroller.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ScrollerFragment">
+    tools:context=".unused.ScrollerFragment">
 
     <ScrollView
         android:layout_width="0dp"
