+++ app/src/main/res/layout/fragment_spinner.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SpinnerFragment">
+    tools:context=".unused.SpinnerFragment">
 
     <Spinner
         android:id="@+id/spinner"
