+++ app/src/main/res/layout/fragment_view.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ViewFragment">
+    tools:context=".fragments.ViewFragment">
 
     <TextView
         android:id="@+id/textName"
